pyproject.toml
[project]
name = "data-viz-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "kaggle>=1.7.4.5",
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "SIM"]
exclude = ["build", "dist", ".venv"]
